chore: move tests after refactoring
